[CMS-327] Upgrade ruby to 2.7 (#12)
- Ignore vendor gems
- Update travis.yml
- Fixes tests
When tests were removed from the suite it was because it appeared that the feature they were testing had been changed since the test were last run (portlets, page creation workflow etc). If the change was obvious then the test was updated to match the new behavior.

In a few places BCMS needed to be modified to pass. Paperclip, in particular, had a significant change around file validation which required the addition of some attachment validations. Paperclip requires the attachment to have at least one validation defined. It also adds file content validation which needed to be disabled.